script updates

+++ scripts/aptlocal.sh
@@ -28,14 +28,19 @@
 # if there is a file called /root/home then use the local mirror
 #
 # set proxy
-echo "http_proxy=http://88.198.57.112:3128" >> /etc/environment
+echo "http_proxy=http://10.5.1.34:3128" >> /etc/environment
 
-#if [ -e "/root/home" ]
-#then 
-   echo "Acquire::http::Proxy \"http://88.198.57.112:3128\";" >> /etc/apt/apt.conf
+if [ -e "/root/home" ]
+then 
+#   echo "Acquire::http::Proxy \"http://10.5.1.34:3128\";" >> /etc/apt/apt.conf
    echo "deb ftp://mirror.hetzner.de/ubuntu/packages precise main restricted universe multiverse" > /etc/apt/sources.list
-#else
-#   echo "deb http://archive.ubuntu.com/ubuntu precise main universe" > /etc/apt/sources.list
-#fi
+   echo "deb ftp://mirror.hetzner.de/ubuntu/packages precise-updates main restricted universe multiverse" >> /etc/apt/sources.list
+else
+   echo "deb http://archive.ubuntu.com/ubuntu/ precise main restricted universe multiverse" > /etc/apt/sources.list
+   echo "deb http://archive.ubuntu.com/ubuntu/ precise-updates main restricted universe multiverse" >> /etc/apt/sources.list
+   echo "deb http://us.archive.ubuntu.com/ubuntu/ precise-security main restricted universe multiverse" >> /etc/apt/sources.list
+fi
+echo "This is sources.list:"
+cat /etc/apt/sources.list
 apt-get update
 apt-get upgrade -y
